backend/CamStyle/requirements.txt
absl-py==0.2.0
apturl==0.5.2
astor==0.6.2
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
feedparser==5.1.3
gast==0.2.0
grpcio==1.12.0
guacamole==0.9.2
html5lib==0.9999999
httplib2==0.9.1
idna==2.0
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==0.23
numpy==1.14.3
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pathlib==1.0.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
protobuf==3.5.2.post1
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.1
python-debian==0.1.27
python-systemd==231
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.10.0
ssh-import-id==5.5
system-service==0.3
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
tqdm==4.23.4
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.2.0
Werkzeug==0.14.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
you-get==0.4.1099
